move trace parser, extract data structures
