+++ src/som/primitives/TimerPrim.java
@@ -29,8 +29,8 @@ import som.vm.constants.Classes;
 import som.vmobjects.SBlock;
 import som.vmobjects.SInvokable;
 import som.vmobjects.SSymbol;
-import tools.concurrency.TraceParser;
 import tools.concurrency.TracingActors.ReplayActor;
+import tools.replay.TraceParser;
 import tools.replay.actors.ActorExecutionTrace;
 import tools.replay.actors.ExternalEventualMessage.ExternalDirectMessage;
 import tools.replay.nodes.TraceContextNode;
