+++ src/tools/concurrency/TracingActors.java
@@ -18,11 +18,12 @@ import som.interpreter.actors.EventualMessage;
 import som.interpreter.actors.EventualMessage.PromiseMessage;
 import som.interpreter.actors.SPromise.STracingPromise;
 import som.vm.VmSettings;
-import tools.concurrency.TraceParser.ExternalMessageRecord;
-import tools.concurrency.TraceParser.ExternalPromiseMessageRecord;
-import tools.concurrency.TraceParser.MessageRecord;
-import tools.concurrency.TraceParser.PromiseMessageRecord;
 import tools.debugger.WebDebugger;
+import tools.replay.TraceParser;
+import tools.replay.TraceParser.ExternalMessageRecord;
+import tools.replay.TraceParser.ExternalPromiseMessageRecord;
+import tools.replay.TraceParser.MessageRecord;
+import tools.replay.TraceParser.PromiseMessageRecord;
 import tools.replay.actors.ExternalMessage;
 import tools.snapshot.SnapshotRecord;
 import tools.snapshot.deserialization.DeserializationBuffer;
